♻️ ⚡ 新規チャットのフックを分離
